release db connection lock so as to not deadlock ci_driver
afaict holding the lock over an await point gives tokio an opportunity
to run a different task - perhaps a different API request that would try
acquring the db connection lock. then it deadlocks; the second task
holds the thread hostage while waiting for the first task to release its
lock, which it can't do due to no thread being available for it to run
on.